requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
sqlalchemy>=2.0.27
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic[email]>=2.7.0
cloudinary>=1.40.0
passlib[bcrypt]>=1.7.4